Move style out of hook
