+++ packages/base/src/features/story/components/SpectaMobileView.tsx
@@ -8,7 +8,7 @@ import {
   DrawerTrigger,
 } from '@/src/shared/components/Drawer';
 import StoryViewerPanel from '../StoryViewerPanel';
-import { getSpectaPresentationStyle } from '../hooks/useStoryMap';
+import { getSpectaPresentationStyle } from '@/src/features/story/utils/spectaPresentation';
 
 const MAIN_ID = 'jp-main-content-panel';
 const SEGMENT_PANEL_ID = 'jgis-story-segment-panel';
